ooinstall: add support for unattended installation
This adds support for unattended installs to ooinstall. Some refactoring was
done to better read in defaults passed in via installer.cfg.yaml and/or
command line parameters as well as separating more of the installer logic
from the UX prompts. Additionially some of the defaults were tweaked
to better support the above