Replace deployment type with product.
This is an attempt to expose more consistent versions to be installed to the
user, and to pave the way for upgrades in the future. Deployment type CLI
option is dropped, you must either specify it in the config file, or
interactively during the install. The more friendly and explicit keys map to
the same old values to be sent to ansible behind the scenes.